send_to_bq/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery

